Rename dirname to parentDirectory
