+++ packages/core/src/cli/logger.ts
@@ -6,7 +6,7 @@ import terminalLink from 'terminal-link'
 import { isRunningTests } from './environment'
 import { LoggerContentToken, LoggerContentType } from './logger/content'
 import { LoggerTarget, NoopLoggerTarget } from './logger/target'
-import { relativizePath } from '../node/path.public'
+import { relativizePath } from '../node/path'
 
 /**
  * We cache the loggers to ensure we only have an
