+++ packages/core/src/cli/project/load/targets/main.test.ts
@@ -2,7 +2,7 @@ import { loadMainTarget } from './main'
 import { describe, test, expect, vi } from 'vitest'
 import { temporary } from '@gestaltjs/testing'
 import { models } from '@gestaltjs/testing'
-import { dirname, basename, joinPath } from '../../../../node/path.public'
+import { parentDirectory, basename, joinPath } from '../../../../node/path'
 import { loadRoutes } from './main/routes'
 import { createRouter } from 'radix3'
 import { Route } from '../../models/targets/main/route'
@@ -26,8 +26,8 @@ describe('loadMainTarget', () => {
 
       // Then
       expect(got.manifestPath).toEqual(manifestPath)
-      expect(got.name).toEqual(basename(dirname(manifestPath)))
-      expect(got.directory).toEqual(dirname(manifestPath))
+      expect(got.name).toEqual(basename(parentDirectory(manifestPath)))
+      expect(got.directory).toEqual(parentDirectory(manifestPath))
       expect(got.router).toBe(router)
     })
   })
