+++ packages/core/src/cli/project/load/targets/main/routes.test.ts
@@ -2,7 +2,7 @@ import { loadRoutes } from './routes'
 import { describe, test, expect, vi } from 'vitest'
 import { temporary } from '@gestaltjs/testing'
 import { models } from '@gestaltjs/testing'
-import { dirname, joinPath } from '../../../../../node/path.public'
+import { parentDirectory, joinPath } from '../../../../../node/path'
 import { writeFile, makeDirectory } from '../../../../../shared/fs'
 import { Route } from '../../../models/targets/main/route'
 
@@ -20,9 +20,9 @@ describe('loadRoutes', () => {
       const postListFilePath = joinPath(tmpDir, 'posts/[post].list.ts')
       const postGetFilePath = joinPath(tmpDir, 'posts/[post].get.js')
 
-      await makeDirectory(dirname(aboutUIFilePath))
-      await makeDirectory(dirname(settingsUIFilePath))
-      await makeDirectory(dirname(postUIFilePath))
+      await makeDirectory(parentDirectory(aboutUIFilePath))
+      await makeDirectory(parentDirectory(settingsUIFilePath))
+      await makeDirectory(parentDirectory(postUIFilePath))
 
       await writeFile(aboutUIFilePath, '')
       await writeFile(settingsUIFilePath, '')
