+++ packages/core/src/cli/tsc.test.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect, vi } from 'vitest'
 import { exec } from './system'
-import { findPathUp, dirname } from '../node/path.public'
+import { findPathUp, parentDirectory } from '../node/path'
 import { run, TSCNotFoundError } from './tsc'
 
 vi.mock('./system')
@@ -12,7 +12,7 @@ describe('run', () => {
     const tscPath = '/test/tsc'
     const dirnamePath = '/gestalt/tsc'
     vi.mocked(findPathUp).mockResolvedValue(tscPath)
-    vi.mocked(dirname).mockReturnValue(dirnamePath)
+    vi.mocked(parentDirectory).mockReturnValue(dirnamePath)
     const args = ['foo']
     const cwd = '/project'
 
@@ -30,7 +30,7 @@ describe('run', () => {
     // Given
     const eslintTSDirectory = '/gestalt/tsc'
     vi.mocked(findPathUp).mockResolvedValue(undefined)
-    vi.mocked(dirname).mockReturnValue(eslintTSDirectory)
+    vi.mocked(parentDirectory).mockReturnValue(eslintTSDirectory)
     const args = ['foo']
     const cwd = '/project'
