+++ packages/core/src/cli/tsc.ts
@@ -1,6 +1,6 @@
 import { Abort } from '../shared/error'
 import { exec } from './system'
-import { findPathUp, dirname } from '../node/path.public'
+import { findPathUp, parentDirectory } from '../node/path'
 import { fileURLToPath } from 'url'
 
 export const TSCNotFoundError = () => {
@@ -8,7 +8,7 @@ export const TSCNotFoundError = () => {
 }
 
 export async function run(args: string[], cwd: string) {
-  const __dirname = dirname(fileURLToPath(import.meta.url))
+  const __dirname = parentDirectory(fileURLToPath(import.meta.url))
   const tscPath = await findPathUp('node_modules/.bin/tsc', { cwd: __dirname })
   if (!tscPath) {
     throw TSCNotFoundError()
